+++ CHANGELOG.md
@@ -21,7 +21,7 @@
 
 #### 🚀  Version 1
 
-- Add compatibility to run on forked PRs (needs public appCode)
+- Add compatibility to run on forked PRs (needs public projectToken)
 - Performance improvement: Run GitHub deployment APIs and Chromatic in parallel
 
 # v1 (Sat Oct 19 2019)